+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/roadrunner/drive/ClawTest.kt
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.opmode.telop
+package org.firstinspires.ftc.teamcode.roadrunner.drive
 
 import com.qualcomm.hardware.sparkfun.SparkFunOTOS
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode
